refactor: relocate security scheme contracts and update references
