+++ src/Host/Configuration/RestfulConfiguration.cs
@@ -1,4 +1,5 @@
 using AuthKit.Plugins.Abstractions;
+using AuthKit.Plugins.Abstractions.Contracts.SecuritySchemes;
 using Host.Plugins;
 using Microsoft.OpenApi.Models;
